requirements.txt
diskcache==5.6.3
llama_cpp_python==0.2.11
numpy==1.26.0
setuptools==68.2.2
typing_extensions==4.8.0
wheel==0.41.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
atlassian-python-api==3.41.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
dataclasses-json==0.6.1
Deprecated==1.2.14
diskcache==5.6.3
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.0
html2markdown==0.1.7
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.311
langsmith==0.0.43
llama_cpp_python==0.2.11
markdownify==0.11.6
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wrapt==1.15.0
yarl==1.9.2
